refactor: folder명 변경
